iOS build not functioning due to dotenv function
line 9 in index.tsx is causing the following error when building on an iOS device:

**iOS Bundling failed
The package at "node_modules\dotenv\lib\main.js" attempted to import the Node standard library module "path".
It failed because the native React runtime does not include the Node standard library.
Learn more: https://docs.expo.dev/workflow/using-libraries/#using-third-party-libraries**

Removing this line fixes this error and allows me to view the pages, however I don't know if this will cause anything in the backend to fail. We need to check that this change won't cause any issues and, if so, proceed with removing the line.